Basic chatpad support on XBox 360 wireless controller
controller_factory.cpp is modified to send chatpad option to
Xbox360WirelessController constructor.
If chatpad option is enabled, LinuxUinput would be instantiated in the
way similar to implementation in chatpad.cpp. But class Chatpad is
tightly coupled with usb communication, the implementation could not be
reused. Defines in chatpad.hpp are reused.
Keymap is slightly altered.
Currently leds on chatpad is not supported.